+++ dailyq/src/main/java/com/knuissant/dailyq/dto/answers/AnswerArchiveUpdateRequest.java
@@ -1,5 +1,4 @@
-package com.knuissant.dailyq.dto;
-
+package com.knuissant.dailyq.dto.answers;
 
 public record AnswerArchiveUpdateRequest(
         String memo,
