+++ dailyq/src/main/java/com/knuissant/dailyq/dto/answers/AnswerCreateRequest.java
@@ -1,4 +1,4 @@
-package com.knuissant.dailyq.dto;
+package com.knuissant.dailyq.dto.answers;
 
 import jakarta.validation.constraints.AssertTrue;
 import jakarta.validation.constraints.NotNull;
