+++ dailyq/src/main/java/com/knuissant/dailyq/dto/answers/AnswerCreateResponse.java
@@ -1,4 +1,4 @@
-package com.knuissant.dailyq.dto;
+package com.knuissant.dailyq.dto.answers;
 
 public record AnswerCreateResponse(
         Long answerId,
