+++ dailyq/src/main/java/com/knuissant/dailyq/dto/answers/AnswerLevelUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.knuissant.dailyq.dto;
+package com.knuissant.dailyq.dto.answers;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
