+++ dailyq/src/main/java/com/knuissant/dailyq/dto/answers/AnswerLevelUpdateResponse.java
@@ -1,4 +1,4 @@
-package com.knuissant.dailyq.dto;
+package com.knuissant.dailyq.dto.answers;
 
 public record AnswerLevelUpdateResponse(
         Long answerId,
