+++ dailyq/src/main/java/com/knuissant/dailyq/dto/answers/AnswerSearchConditionRequest.java
@@ -1,4 +1,4 @@
-package com.knuissant.dailyq.dto;
+package com.knuissant.dailyq.dto.answers;
 
 import java.time.LocalDate;
