+++ dailyq/src/main/java/com/knuissant/dailyq/dto/users/UserJobsInfoResponse.java
@@ -1,4 +1,4 @@
-package com.knuissant.dailyq.dto;
+package com.knuissant.dailyq.dto.users;
 
 import java.util.List;
