+++ dailyq/src/main/java/com/knuissant/dailyq/dto/users/UserJobsUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.knuissant.dailyq.dto;
+package com.knuissant.dailyq.dto.users;
 
 import jakarta.validation.constraints.NotNull;
