+++ dailyq/src/main/java/com/knuissant/dailyq/dto/users/UserPreferencesResponse.java
@@ -1,4 +1,4 @@
-package com.knuissant.dailyq.dto;
+package com.knuissant.dailyq.dto.users;
 
 import java.time.LocalTime;
