+++ dailyq/src/main/java/com/knuissant/dailyq/dto/users/UserPreferencesUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.knuissant.dailyq.dto;
+package com.knuissant.dailyq.dto.users;
 
 import com.knuissant.dailyq.domain.questions.QuestionMode;
 import com.knuissant.dailyq.domain.users.UserResponseType;
