+++ dailyq/src/main/java/com/knuissant/dailyq/service/AnswerService.java
@@ -24,20 +24,20 @@ import com.knuissant.dailyq.domain.feedbacks.FeedbackStatus;
 import com.knuissant.dailyq.domain.jobs.Job;
 import com.knuissant.dailyq.domain.questions.Question;
 import com.knuissant.dailyq.domain.users.User;
-import com.knuissant.dailyq.dto.AnswerArchiveUpdateRequest;
-import com.knuissant.dailyq.dto.AnswerArchiveUpdateResponse;
-import com.knuissant.dailyq.dto.AnswerDetailResponse;
-import com.knuissant.dailyq.dto.AnswerListResponse.CursorResult;
-import com.knuissant.dailyq.dto.AnswerListResponse.Summary;
-import com.knuissant.dailyq.dto.AnswerSearchConditionRequest;   
+import com.knuissant.dailyq.dto.answers.AnswerArchiveUpdateRequest;
+import com.knuissant.dailyq.dto.answers.AnswerArchiveUpdateResponse;
+import com.knuissant.dailyq.dto.answers.AnswerCreateRequest;
+import com.knuissant.dailyq.dto.answers.AnswerCreateResponse;
+import com.knuissant.dailyq.dto.answers.AnswerDetailResponse;
+import com.knuissant.dailyq.dto.answers.AnswerLevelUpdateRequest;
+import com.knuissant.dailyq.dto.answers.AnswerLevelUpdateResponse;
+import com.knuissant.dailyq.dto.answers.AnswerSearchConditionRequest;
+import com.knuissant.dailyq.dto.answers.AnswerListResponse.CursorResult;
+import com.knuissant.dailyq.dto.answers.AnswerListResponse.Summary;
 import com.knuissant.dailyq.exception.BusinessException;
 import com.knuissant.dailyq.exception.ErrorCode;
 import com.knuissant.dailyq.repository.AnswerRepository;
 import com.knuissant.dailyq.repository.FeedbackRepository;
-import com.knuissant.dailyq.dto.AnswerCreateRequest;
-import com.knuissant.dailyq.dto.AnswerCreateResponse;
-import com.knuissant.dailyq.dto.AnswerLevelUpdateRequest;
-import com.knuissant.dailyq.dto.AnswerLevelUpdateResponse;
 import com.knuissant.dailyq.repository.QuestionRepository;
 import com.knuissant.dailyq.repository.UserRepository;
