+++ dailyq/src/main/java/com/knuissant/dailyq/service/JobService.java
@@ -6,7 +6,7 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.knuissant.dailyq.dto.JobResponse;
+import com.knuissant.dailyq.dto.users.JobResponse;
 import com.knuissant.dailyq.exception.BusinessException;
 import com.knuissant.dailyq.exception.ErrorCode;
 import com.knuissant.dailyq.repository.JobRepository;
