+++ dailyq/src/main/java/com/knuissant/dailyq/service/UserPreferencesService.java
@@ -8,9 +8,9 @@ import com.knuissant.dailyq.domain.questions.QuestionMode;
 import com.knuissant.dailyq.domain.users.User;
 import com.knuissant.dailyq.domain.users.UserPreferences;
 import com.knuissant.dailyq.domain.users.UserResponseType;
-import com.knuissant.dailyq.dto.UserJobsUpdateRequest;
-import com.knuissant.dailyq.dto.UserPreferencesResponse;
-import com.knuissant.dailyq.dto.UserPreferencesUpdateRequest;
+import com.knuissant.dailyq.dto.users.UserJobsUpdateRequest;
+import com.knuissant.dailyq.dto.users.UserPreferencesResponse;
+import com.knuissant.dailyq.dto.users.UserPreferencesUpdateRequest;
 import com.knuissant.dailyq.exception.BusinessException;
 import com.knuissant.dailyq.exception.ErrorCode;
 import com.knuissant.dailyq.repository.JobRepository;
